Move ffio code into io
